move magik-farm to a separate folder and include staking, pool2
