+++ projects/magik-farm/config/vault/aurora_pools.js
