+++ projects/magik-farm/config/vault/avalanche_pools.js
