+++ projects/magik-farm/config/vault/cronos_pools.js
