+++ projects/magik-farm/config/vault/fantom_pools.js
